add tm_calibration package
based on some scripts in `tmr_description`. This package mimics the
ur_calibration API, in the sense that it exposes a executable which
fetches the factory calibration from a robot and creates a YAML file
which can be fed to a tmX_upload.launch to ensure that a correct URDF is
put on the parameter server.

The main difference with ur_calibration is that the executable in this
package requires that the TM ROS driver is currently running, as it
relies on services exposed by the driver in order to fetch the factory
calibration.