update tm{12,5}_description packages to accept a kinematics file
I.e., the type of kinematics file created by tm_calibration. This
involves editing the tmX_upload.launch and related
tmX{_robot}.urdf.xacro files.

Since the visual/collision meshes are not oriented in the same way as
the kinematics tree, this required adding visual/collision origins for
the majority of links.

Once the description packages are merged to one, duplication can be
reduced by a lot.