Add kinematics support (#5)
* add tm_calibration package

based on some scripts in `tmr_description`. This package mimics the
ur_calibration API, in the sense that it exposes a executable which
fetches the factory calibration from a robot and creates a YAML file
which can be fed to a tmX_upload.launch to ensure that a correct URDF is
put on the parameter server.

* update tm{12,5}_description packages to accept a kinematics file

I.e., the type of kinematics file created by tm_calibration. This
involves editing the tmX_upload.launch and related
tmX{_robot}.urdf.xacro files.

Since the visual/collision meshes are not oriented in the same way as
the kinematics tree, this required adding visual/collision origins for
the majority of links.

Once the description packages are merged to one, duplication can be
reduced by a lot.

* update tm_driver top-level launch files (support kinematics)

* expose base_frame and tool_frame args in tmX_bringup.launch

Co-authored-by: Leo Ghafari <leo@ascent.ai>